Move several client model implementations up to the base plugin
