+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp.ClientModel/test/Mocks/MockTypeFactory.cs
@@ -3,12 +3,13 @@
 
 using System;
 using System.Collections.Generic;
+using Microsoft.Generator.CSharp.ClientModel;
 using Microsoft.Generator.CSharp.Input;
 using Microsoft.Generator.CSharp.Providers;
 
-namespace Microsoft.Generator.CSharp.Tests
+namespace Microsoft.Generator.CSharp.ClientModel.Tests
 {
-    internal class MockTypeFactory : TypeFactory
+    internal class MockTypeFactory : ScmTypeFactory
     {
         public override MethodProviderCollection? CreateMethodProviders(InputOperation operation, TypeProvider enclosingType)
         {
@@ -25,6 +26,13 @@ namespace Microsoft.Generator.CSharp.Tests
             throw new NotImplementedException();
         }
 
+        public override CSharpType MatchConditionsType() => typeof(int);
+
+        public override CSharpType PageResponseType() => typeof(int);
+
+        public override CSharpType RequestConditionsType() => typeof(int);
+
+        public override CSharpType TokenCredentialType() => typeof(int);
         public override CSharpType ListInitializationType => new CSharpType(typeof(List<>), arguments: typeof(int));
         public override CSharpType DictionaryInitializationType => new CSharpType(typeof(Dictionary<,>), arguments: [typeof(string), typeof(int)]);
     }
