+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp/src/OutputTypes/TypeFactory.cs
@@ -3,6 +3,9 @@
 
 using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Net;
 using Microsoft.Generator.CSharp.Input;
 using Microsoft.Generator.CSharp.Providers;
 
@@ -15,14 +18,59 @@ namespace Microsoft.Generator.CSharp
         /// </summary>
         /// <param name="input">The <see cref="InputType"/> to convert.</param>
         /// <returns>An instance of <see cref="CSharpType"/>.</returns>
-        public abstract CSharpType CreateCSharpType(InputType input);
+        public virtual CSharpType CreateCSharpType(InputType inputType) => inputType switch
+        {
+            InputLiteralType literalType => CSharpType.FromLiteral(CreateCSharpType(literalType.ValueType), literalType.Value),
+            InputUnionType unionType => CSharpType.FromUnion(unionType.VariantTypes.Select(CreateCSharpType).ToArray(), unionType.IsNullable),
+            InputListType { IsEmbeddingsVector: true } listType => new CSharpType(typeof(ReadOnlyMemory<>), listType.IsNullable, CreateCSharpType(listType.ElementType)),
+            InputListType listType => new CSharpType(typeof(IList<>), listType.IsNullable, CreateCSharpType(listType.ElementType)),
+            InputDictionaryType dictionaryType => new CSharpType(typeof(IDictionary<,>), inputType.IsNullable, typeof(string), CreateCSharpType(dictionaryType.ValueType)),
+            InputEnumType enumType => CodeModelPlugin.Instance.OutputLibrary.EnumMappings.TryGetValue(enumType, out var provider)
+                ? provider.Type.WithNullable(inputType.IsNullable)
+                : throw new InvalidOperationException($"No {nameof(EnumProvider)} has been created for `{enumType.Name}` {nameof(InputEnumType)}."),
+            InputModelType model => CodeModelPlugin.Instance.OutputLibrary.ModelMappings.TryGetValue(model, out var provider)
+                ? provider.Type.WithNullable(inputType.IsNullable)
+                : new CSharpType(typeof(object), model.IsNullable).WithNullable(inputType.IsNullable),
+            InputPrimitiveType primitiveType => primitiveType.Kind switch
+            {
+                InputPrimitiveTypeKind.Boolean => new CSharpType(typeof(bool), inputType.IsNullable),
+                InputPrimitiveTypeKind.Bytes => new CSharpType(typeof(BinaryData), inputType.IsNullable),
+                InputPrimitiveTypeKind.ContentType => new CSharpType(typeof(string), inputType.IsNullable),
+                InputPrimitiveTypeKind.PlainDate => new CSharpType(typeof(DateTimeOffset), inputType.IsNullable),
+                InputPrimitiveTypeKind.Decimal => new CSharpType(typeof(decimal), inputType.IsNullable),
+                InputPrimitiveTypeKind.Decimal128 => new CSharpType(typeof(decimal), inputType.IsNullable),
+                InputPrimitiveTypeKind.PlainTime => new CSharpType(typeof(TimeSpan), inputType.IsNullable),
+                InputPrimitiveTypeKind.Float32 => new CSharpType(typeof(float), inputType.IsNullable),
+                InputPrimitiveTypeKind.Float64 => new CSharpType(typeof(double), inputType.IsNullable),
+                InputPrimitiveTypeKind.Float128 => new CSharpType(typeof(decimal), inputType.IsNullable),
+                InputPrimitiveTypeKind.Guid or InputPrimitiveTypeKind.Uuid => new CSharpType(typeof(Guid), inputType.IsNullable),
+                InputPrimitiveTypeKind.Int8 => new CSharpType(typeof(sbyte), inputType.IsNullable),
+                InputPrimitiveTypeKind.UInt8 => new CSharpType(typeof(byte), inputType.IsNullable),
+                InputPrimitiveTypeKind.Int32 => new CSharpType(typeof(int), inputType.IsNullable),
+                InputPrimitiveTypeKind.Int64 => new CSharpType(typeof(long), inputType.IsNullable),
+                InputPrimitiveTypeKind.SafeInt => new CSharpType(typeof(long), inputType.IsNullable),
+                InputPrimitiveTypeKind.Integer => new CSharpType(typeof(long), inputType.IsNullable), // in typespec, integer is the base type of int related types, see type relation: https://typespec.io/docs/language-basics/type-relations
+                InputPrimitiveTypeKind.Float => new CSharpType(typeof(double), inputType.IsNullable), // in typespec, float is the base type of float32 and float64, see type relation: https://typespec.io/docs/language-basics/type-relations
+                InputPrimitiveTypeKind.Numeric => new CSharpType(typeof(double), inputType.IsNullable), // in typespec, numeric is the base type of number types, see type relation: https://typespec.io/docs/language-basics/type-relations
+                InputPrimitiveTypeKind.IPAddress => new CSharpType(typeof(IPAddress), inputType.IsNullable),
+                InputPrimitiveTypeKind.Stream => new CSharpType(typeof(Stream), inputType.IsNullable),
+                InputPrimitiveTypeKind.String => new CSharpType(typeof(string), inputType.IsNullable),
+                InputPrimitiveTypeKind.Uri or InputPrimitiveTypeKind.Url => new CSharpType(typeof(Uri), inputType.IsNullable),
+                InputPrimitiveTypeKind.Char => new CSharpType(typeof(char), inputType.IsNullable),
+                InputPrimitiveTypeKind.Any => new CSharpType(typeof(BinaryData), inputType.IsNullable),
+                _ => new CSharpType(typeof(object), inputType.IsNullable),
+            },
+            InputDateTimeType dateTimeType => new CSharpType(typeof(DateTimeOffset), inputType.IsNullable),
+            InputDurationType durationType => new CSharpType(typeof(TimeSpan), inputType.IsNullable),
+            _ => throw new Exception("Unknown type")
+        };
 
         /// <summary>
         /// Factory method for creating a <see cref="ParameterProvider"/> based on an input parameter <paramref name="parameter"/>.
         /// </summary>
         /// <param name="parameter">The <see cref="InputParameter"/> to convert.</param>
         /// <returns>An instance of <see cref="ParameterProvider"/>.</returns>
-        public abstract ParameterProvider CreateCSharpParam(InputParameter parameter);
+        public virtual ParameterProvider CreateCSharpParam(InputParameter parameter) => new ParameterProvider(parameter);
 
         /// <summary>
         /// Factory method for creating a <see cref="MethodProviderCollection"/> based on an input operation <paramref name="operation"/>.
@@ -78,11 +126,6 @@ namespace Microsoft.Generator.CSharp
             },
             _ => SerializationFormat.Default
         };
-
-        public abstract CSharpType MatchConditionsType();
-        public abstract CSharpType RequestConditionsType();
-        public abstract CSharpType TokenCredentialType();
-        public abstract CSharpType PageResponseType();
 
         /// <summary>
         /// The initialization type of list properties. This type should implement both <see cref="IList{T}"/> and <see cref="IReadOnlyList{T}"/>.
