+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp/test/TypeFactoryTests.cs
@@ -58,27 +58,6 @@ namespace Microsoft.Generator.CSharp.Tests
 
         internal class CustomTypeFactory : TypeFactory
         {
-
-            public override CSharpType RequestConditionsType()
-            {
-                throw new NotImplementedException();
-            }
-
-            public override CSharpType TokenCredentialType()
-            {
-                throw new NotImplementedException();
-            }
-
-            public override CSharpType MatchConditionsType()
-            {
-                throw new NotImplementedException();
-            }
-
-            public override CSharpType PageResponseType()
-            {
-                throw new NotImplementedException();
-            }
-
             public override ParameterProvider CreateCSharpParam(InputParameter parameter)
             {
                 throw new NotImplementedException();
